Move crate::map::core to crate::inner
